webpanel-arduino: add webpanel for Arduino
This patch adds the arduino luci webpanel. In this patch I do not remove
the linino webpanel due to maintenability problems, but I just disable
it from the default configuration.

The master branch is the main branch of the linino distribution and it
uses the linino webpanel. In order to avoid git conflicts while rebasing
arduino branch I did not remove the linino webpanel.

Signed-off-by: Federico Vaga <federicov@linino.org>